### 2019-04-29 f6a1714
bug fix to allow matrix argument for data
