### 2019-04-29 c9a1c2b
bug fix to allow data as matrix
