### 2019-04-29 e28006b
bug fix for data as matrix
