### 2019-04-29 #1415 @katrinekirkeby
Am I allowed to change the Version, if i found some small errors in the code?
