### 2019-04-29 9d1e85c
bug fix for data as matrix
