+++ R/increase_order1.R
@@ -126,6 +126,8 @@ increase_order1 <- function(tch_cliq, data, ...){
     stop("data must be a data frame or a matrix.")
   }
 
+  data <- as.data.frame(data)
+
   if (! all(sapply(data, function(x){
     is.character(x) | is.factor(x)
   }
