+++ R/k_tcherry_p_lookahead.R
@@ -95,6 +95,8 @@ k_tcherry_p_lookahead <- function(data, k, p, ...){
     stop("data must be a data frame or a matrix.")
   }
 
+  data <- as.data.frame(data)
+
   if (! all(sapply(data, function(x){
     is.character(x) | is.factor(x)
   }
