+++ R/increase_order2.R
@@ -16,6 +16,8 @@ increase_order2 <- function(tch_cliq, data, ...){
     stop("data must be a data frame or a matrix.")
   }
 
+  data <- as.data.frame(data)
+
   if (! all(sapply(data, function(x){
     is.character(x) | is.factor(x)
   }
