+++ tests/testthat/testincrease_order2.R
@@ -20,6 +20,8 @@ data <- data.frame("var1" = as.character(var1),
                    "var6" = as.character(var6),
                    "var7" = as.character(var7))
 
+data_mat <- as.matrix(data)
+
 ChowLiu_cliques <- list(c("var1", "var5"),
                         c("var2", "var5"),
                         c("var3", "var5"),
@@ -27,10 +29,11 @@ ChowLiu_cliques <- list(c("var1", "var5"),
                         c("var4", "var6"),
                         c("var5", "var6"))
 
-
-
 tcherry_ord_inc <- increase_order2(ChowLiu_cliques, data,
                                               smooth = 0.1)
+tcherry_ord_inc2 <- increase_order2(ChowLiu_cliques, data_mat,
+                                   smooth = 0.1)
+
 adj_mat <- matrix(c(0, 0, 1, 1, 1, 1, 0,
                     0, 0, 1, 0, 1, 0, 1,
                     1, 1, 0, 0, 1, 0, 1,
@@ -38,6 +41,7 @@ adj_mat <- matrix(c(0, 0, 1, 1, 1, 1, 0,
                     1, 1, 1, 0, 0, 1, 0,
                     1, 0, 0, 1, 1, 0, 0,
                     0, 1, 1, 0, 0, 0, 0), nrow = 7)
+
 colnames(adj_mat) <- rownames(adj_mat) <- names(data)
 
 cliques <- list(c("var2", "var3", "var5"),
@@ -58,6 +62,13 @@ test_that("results are corrects", {
   expect_true(compare::compare(tcherry_ord_inc$cliques, cliques,
                                ignoreOrder = TRUE)$result)
   expect_true(compare::compare(tcherry_ord_inc$separators, seps,
+                               ignoreOrder = TRUE)$result)
+
+  expect_equal(tcherry_ord_inc2$adj_matrix, adj_mat)
+  expect_equal(tcherry_ord_inc2$n_edges, 11)
+  expect_true(compare::compare(tcherry_ord_inc2$cliques, cliques,
+                               ignoreOrder = TRUE)$result)
+  expect_true(compare::compare(tcherry_ord_inc2$separators, seps,
                                ignoreOrder = TRUE)$result)
 })
 
