### 2019-05-09 9713ed2
update documentation to k'th order
