+++ man/diff_edges_tch.Rd
@@ -14,7 +14,8 @@ The number of edges for which the two graphs differ.
 }
 \description{
 Determines the number of differing edges for two t-cherry
-trees over the same universe represented by adjacency matrices,
+trees over the same universe and of the same order represented by
+adjacency matrices,
 i.e. the number of edges in each graph minus the number of edges
 common to both t-cherry trees.
 }
