### 2019-05-15 17d0ccf
Fixing layout, grammer and spelling errors. Adding error messages and test.
