### 2019-05-29 #1480 @whedon
[REVIEW]: tcherry: Learning the structure of tcherry trees
**Submitting author:** @katrinekirkeby (<a href="http://orcid.org/0000-0001-6065-1583">Katrine Kirkeby</a>)
**Repository:** <a href="https://github.com/nvihrs14/tcherry" target ="_blank">https://github.com/nvihrs14/tcherry</a>
**Version:** 0.1.1
