### 2019-05-29 #1415 @jasonclark
@kellieotto The review is open at https://github.com/openjournals/joss-reviews/issues/1480. The checklist there gives you a framework to follow. If you have any questions, let me know. Thanks again for your time and expertise.
