### 2019-05-29 #1415 @kyleniemeyer
@jasonclark can you try to find a second reviewer for the submission? We prefer to have at least two (even though JOSS hasn't consistently applied that expectation up until now). If it is very challenging to find another reviewer, then it's ok for the editor to serve as the second reviewer. Thanks!
