### 2019-06-12 nvihrs14/tcherry#3 @kellieotto
Missing package version
This issue is part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/1480) for `tcherry`.

One criteria is that the GitHub version matches the package version. Currently, the repository shows no release version for the software. Please [tag the package release](https://help.github.com/en/articles/creating-releases) once you are done with the other revisions.
