### 2019-06-12 nvihrs14/tcherry#4 @kellieotto
Software paper revisions
This issue is part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/1480) for `tcherry`.

**Statement of need:** I feel that the paper is unclear on A) what kind of problem the software solves (e.g. finding relationships between variables in some certain kind of data) and B) who the audience for this package is (e.g. biologists?). Please clarify these points.

**References:** not all citations have a corresponding DOI. Please add the DOI to every reference that has has one.

**Proofread:** once you have revised the content, please check the grammar for verb tense. I am happy to help with this if you'd like!
