### 2019-06-12 nvihrs14/tcherry#5 @kellieotto
Installation dependencies
This issue is part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/1480) for `tcherry`.

I had some issues installing the package and building the vignettes due to the dependencies.

**Bioconductor packages:** I would note in your install instructions that the `graph` and `RBGL` packages come from Bioconductor, and that you might need to install them through BiocManager. At first, I tried to install them with `install.packages` and the error message was not helpful.

**Dependencies in the vignette:** I was unable to build the vignette due to missing packages `gRain` and `Rgraphviz`. Either add these to the dependencies, or better yet, remove them from the vignette if they are not essential.
