### 2019-06-12 #1480 @kellieotto
@jasonclark @katrinekirkeby The package looks good on the whole but needs minor revisions. I've detailed them in the issues referenced above. Please tag me when they're done and I will take another look.
