+++ paper.bib
@@ -17,6 +17,7 @@
     number = {10},
     pages = {1--26},
     Url = {http://www.jstatsoft.org/v46/i10/},
+    DOI={10.18637/jss.v046.i10}
   }
   
   @article{ChowLiu,
@@ -44,7 +45,8 @@
   volume={633},
   pages={39-56},
   year={2010},
-  publisher={Springer}
+  publisher={Springer},
+  DOI={10.1007/978-3-642-03735-1_3}
 }
 
 
@@ -54,7 +56,8 @@
   journal={Optim Eng},
   volume={14},
   pages={503-518},
-  year={2013}
+  year={2013},
+  DOI={10.1007/s11081-013-9232-8}
 }
 
 @Book{Proulx,
@@ -70,7 +73,8 @@
   journal={Forensic Science International: Genetics},
   volume={37},
   pages={29-36},
-  year={2018}
+  year={2018},
+  DOI={10.1016/j.fsigen.2018.07.014}
   }
 
 @misc{UCI ,
