+++ paper.bib
@@ -8,6 +8,19 @@
   DOI = {10.1007/s10479-010-0814-y}
 }
 
+@article{patternrec,
+	author = {Tamas Szantai and Edith Kovacs},
+	title = {Application Of t-Cherry Junction Trees in Pattern Recognition},
+	journal = {BRAIN. Broad Research in Artificial Intelligence and Neuroscience},
+	volume = {1},
+	number = {0},
+	year = {2010},
+	keywords = {pattern recognition, probabilistic modeling, h-uniform hyper- tree, t-cherry junction tree},
+	issn = {2067-3957},
+	url = {https://www.edusoft.ro/brain/index.php/brain/article/view/103},
+	pages = {40--45}
+}
+
 @article{gRain,
     title = {Graphical Independence Networks with the {gRain} Package for {R}},
     author = {S{\o}ren H{\o}jsgaard},
@@ -19,7 +32,7 @@
     Url = {http://www.jstatsoft.org/v46/i10/},
     DOI={10.18637/jss.v046.i10}
   }
-  
+
   @article{ChowLiu,
   title={Approximating Discrete Probability Distributions with Dependence Trees},
   author={C. K. Chow and C. N. Liu},
