### 2019-06-14 8d59c7e
update: dependencies and installation guide
