### 2019-06-14 #1480 @katrinekirkeby
@kellieotto Thank you for reviewing this packages.  We have now tried to solved the missing issues, but regarding the issue about gramma we are not quite sure what you mean since we are not native english speakers, could you please clarify what the issues are
