### 2019-06-14 #1415 @katrinekirkeby
> > Hi, @katrinekirkeby. Things look in order. I'm going to look for a reviewer and copy you. Thanks for the submission!
>
> @katrinekirkeby I did see that the description, topics, and website are empty on the repo landing page. Could you add some info there? Topics will help with searching/discoverability. An abridged description from the README will help introduce the software. If you don't have a website, no worries. The other two bits of info will provide the context as someone lands on the repo homepage. Thanks!

Hi, @jasonclark. We have now added a description and topics to the repo landing page.
