### 2019-06-14 nvihrs14/tcherry#4 @kellieotto
Very minor revisions:

To do this _exact_ --> To do this _exactly_
This is in most cases _to_ time-consuming --> _too_ time-consuming
solve _klassification_ problems with categorical variables --> _classification_
