### 2019-06-14 #1480 @kellieotto
Thanks for making those changes so quickly @katrinekirkeby. Everything looks good now. I made some very minor grammar edits to the paper in the issue here: https://github.com/nvihrs14/tcherry/issues/4
