### 2019-06-17 619c96f
Update paper.md
