### 2019-06-17 #1480 @katrinekirkeby
@kellieotto I have now fixed the grammar edits
