### 2019-06-18 #1415 @arfon
> @jasonclark Sorry about the delay - yup, can do it sometime this week.

@dhhagan - I've added a checklist to the review over in #1480 for you.

@jasonclark - can this pre-review issue be closed now?
