### 2019-06-25 #1480 @dhhagan
@arfon For some reason, it keeps saying that it can't find the invitation and won't let me accept.
