### 2019-06-26 14713c1
Update README.md
Just made a few spelling changes and made functions easier to identify by highlighting them as `code`.
