### 2019-06-26 #1480 @arfon
@dhhagan - I'm so sorry for the trouble we've caused you here. You've actually helped me uncover a bug in our infrastructure. This should be fixed in https://github.com/openjournals/joss/commit/16230e01a4fdafc7a9df4f894b8f4a0dc46909b7 but you will have to accept the invite (hopefully for the last time!) again.
