### 2019-06-26 #1480 @arfon
> Hey @arfon and @katrinekirkeby I have completed most of the review - I opened up a couple of issues on the github repository - if you can address those, I will approve ASAP!

Excellent, thanks @dhhagan. @katrinekirkeby - please take a look at these issues and let us know when you have done so.
