### 2019-06-26 #1480 @katrinekirkeby
> Hey @arfon and @katrinekirkeby I have completed most of the review - I opened up a couple of issues on the github repository - if you can address those, I will approve ASAP!

Hello @dhhagan regarding the issue about the names, it is a choice we made because we think the name should say what the function is made for. Regarding the issue about contributions we have stated in the readme that the packages is a part of our Masters thesis, and because of that we don’t will be adding further functions, but we will answer questions over mail
