### 2019-07-14 fdf0f3f
add contributing file
