### 2019-07-14 a1c1086
changing function names
