+++ man/compute_BIC_junction_tree.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/BIC_junction_tree.R
-\name{BIC_junction_tree}
-\alias{BIC_junction_tree}
+\name{compute_BIC_junction_tree}
+\alias{compute_BIC_junction_tree}
 \title{Calculates the BIC value}
 \usage{
-BIC_junction_tree(cliques, separators, data, base_log = 2, ...)
+compute_BIC_junction_tree(cliques, separators, data, base_log = 2, ...)
 }
 \arguments{
 \item{cliques}{A list containing the cliques of the junction tree.}
@@ -63,7 +63,7 @@ separators <- list(c("var1"),
                    c("var5", "var6"))
 
 # smooth is used to deal with zero probabilities.
-BIC_junction_tree(cliques, separators, data, smooth = 0.1)
+compute_BIC_junction_tree(cliques, separators, data, smooth = 0.1)
 }
 \seealso{
 \code{\link{loglikelihood}} for calculating the

[…]
