### 2019-07-16 #1480 @katrinekirkeby
> I've just emailed the submitting author to ask them to return to this submission soon.

Sorry for the late answer, somehow the emails ended up in my spam box. I'm not quite sure how to make an archive of the software in Zenodo/figshare/other service, but I have baked a new release of the software and made a new release on GitHub where this number is under the tags version: e472a7a
