### 2019-07-17 #1480 @mikldk
@katrinekirkeby See e.g. <https://genr.eu/wp/cite/> (or some other guide) on how to do that. It is only a few steps.
