### 2019-07-18 #1480 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
